Adding joining file code. Add file via upload.
I added four files with the code to join the event, transaction, metadata, and parquet file to the train file. The code is given individually for each task if only one is needed. There is a combined one also named "join_all.py" which has code to perform all three tasks together, and there are also two lines of code at the last which you want to comment out; they are basically to make a new parquet file for your edited or joined data. Comments are provided in the code for help.😊